+++ src/pages/sign-up/ui/personal-information/variants.ts
@@ -1,0 +1,3 @@
+import { cva } from 'class-variance-authority'
+
+export const radioFieldWrapper = cva('flex gap-[30px]')
